+++ packages/rebuild-operation/src/yarnBuild.ts
@@ -1,6 +1,6 @@
 import { spawnSync } from "child_process";
 import { buildFolderName } from "filename-conventions";
-import { getLastFolder, path } from "fs-util";
+import { fs, getLastFolder, path } from "fs-util";
 import { log } from "log";
 import { minifyBuild } from "minify-build";
 import { executeCommandQuietUnlessFail } from "./executeCommandQuietUnlessFail";
@@ -18,7 +18,7 @@ export const yarnBuild = async (
     rmFirst?: boolean;
   }
 ): Promise<Success> => {
-  if (config?.rmFirst) {
+  if (config?.rmFirst && fs.existsSync(path.join(operationBasePath, "build"))) {
     const removed = spawnSync(`rm -rf build`, {
       cwd: operationBasePath,
       encoding: "utf8",
@@ -26,13 +26,18 @@ export const yarnBuild = async (
       shell: true,
     });
   }
+
+  if (!fs.existsSync(operationBasePath)) {
+    return false;
+  }
+
   // TODO:
   // this should only happen conditionally but for now we'll always try
   // later we can add the option to not build if any dependant builds break because of our code changes, but this is quite complex.
   const success = executeCommandQuietUnlessFail({
     command: "yarn build",
     cwd: operationBasePath,
-    description: "Compiling source",
+    description: `Compiling source ${operationBasePath}`,
   });
 
   if (!success) {
@@ -40,7 +45,7 @@ export const yarnBuild = async (
       type: "error",
     });
     return false;
-  }
+  } //
 
   const minified = await minifyBuild({
     buildFolderPath: path.join(operationBasePath, buildFolderName),
